+++ src/lib/Endpoint/Balloon.php
@@ -100,6 +100,6 @@ class Balloon extends AbstractRest
             throw new Exception\ObjectNotFound('no object found with filter '.$filter);
         }
 
-        return $this->build(array_shift($data));
+        return $this->build(array_shift($data), $filter);
     }
 }