+++ src/lib/Endpoint/Pdo.php
@@ -42,7 +42,7 @@ class Pdo extends AbstractSqlDatabase
      */
     public function getAll(?array $query = null): Generator
     {
-        $filter = $this->transformQuery($query);
+        list($filter, $values) = $this->transformQuery($query);
         $this->logGetAll($filter);
 
         if ($filter === null) {
@@ -52,7 +52,7 @@ class Pdo extends AbstractSqlDatabase
         }
 
         try {
-            $result = $this->socket->select($sql);
+            $result = $this->socket->prepareValues($sql, $values);
         } catch (InvalidQuery $e) {
             $this->logger->error('failed to fetch resources from endpoint', [
                 'class' => get_class($this),
@@ -76,19 +76,21 @@ class Pdo extends AbstractSqlDatabase
      */
     public function getOne(array $object, array $attributes = []): EndpointObjectInterface
     {
-        $filter = $this->transformQuery($this->getFilterOne($object));
+        list($filter, $values) = $query = $this->transformQuery($this->getFilterOne($object));
         $this->logGetOne($filter);
         $sql = 'SELECT * FROM '.$this->table.' WHERE '.$filter;
-        $result = $this->socket->select($sql);
+        $result = $this->socket->prepareValues($sql, $values);
+        $row = $result->fetch(\PDO::FETCH_ASSOC);
+        $seccond = $result->fetch(\PDO::FETCH_ASSOC);
 
-        if ($result->num_rows > 1) {
+        if ($seccond !== false) {
             throw new Exception\ObjectMultipleFound('found more than one object with filter '.$filter);
         }
-        if ($result->num_rows === 0) {
+        if ($row === false) {
             throw new Exception\ObjectNotFound('no object found with filter '.$filter);
         }
 
-        return $this->build($result->fetch_assoc());
+        return $this->build($row, $query);
     }
 
     /**